Add separate module for providing OkHttp interceptors
